added functionality to order page
added tabs and display the items by category using the mysql data. Tabs
are really weird and frustruating to use, I created two more html
(tabtester.html and tabtester2.html) if anyone wanna practice with it
and modify our tab. Also, there's weird padding problem with the tabs,
Ted do you know what's going on? Also, is the check out list on the
right side purely display, or is there function I can use to retrieve
the list and its info? If it is purely for display, then I need to
create functions to remember these items for check out and store in
database.