fixed the error message
i notice it's an earlier version with a lot of things undone ..someone
should remember to get the most updated version before editing and
upload it. when i finish working on a page, i always sync it, so if u
see that make sure to grab the latest version in filezilla, thanks! This
isn't the first time, and I don't wanna keep doing maintenance.